chore: moved internal functions
