+++ packages/core/src/plugins/index.ts
@@ -1,3 +1,2 @@
 export * from './load';
-export * from './read';
-export * from './pack';
+export * from './bundle';
