--- packages/core/src/plugins/unpack.ts
@@ -1,6 +1,0 @@
-import AdmZip from 'adm-zip';
-
-export async function unpack(location: string) {
-  const zip = new AdmZip(`${location}.zop`);
-  zip.extractAllTo(location);
-}
